requirements.txt
# Web framework
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
click==8.1.7
itsdangerous==2.2.0
MarkupSafe==2.1.5

# Server
gunicorn==22.0.0

# Env vars
python-dotenv==1.0.1

# Cloudinary (uploads)
cloudinary==1.41.0
